fix: the icon feature problems
